[DotNetCore] Fix duplicate target frameworks displayed
Fixed bug #56782 - Target Framework dropdown shows duplicate values
https://bugzilla.xamarin.com/show_bug.cgi?id=56782

If .NET Core SDKs 1.0.4 and 1.0.3 are installed then there are
multiple runtimes with the same major and minor verisons. This then
resulted in 1.1, 1.1, 1.0, and 1.0 being displayed as target
frameworks in the project template wizard and the project options.
These duplicate values are no longer displayed.