[Packaging] Fix being unable to load nuproj created by Visual Studio
Fixed bug #56701 - Windows created nuspec can not be loaded on a mac
https://bugzilla.xamarin.com/show_bug.cgi?id=56701

Visual Studio on Windows currently creates a .nuproj with no target
framework version which resulted in 4.0 being used by default. The
packaging project was explicitly checking for 4.5 or above. So the
packaging project now indicates it supports any target framework.