[DotNetCore] Do not enable project extension for non Sdk projects
The .NET Core project extension was being enabled for NuGet packaging
projects (.nuproj) when the NuGet.Build.Packaging 0.1.276 was
installed. This version of NuGet.Build.Packaging has changed and
defines a <TargetFramework> property similar to .NET Core projects
so the project extension was incorrectly being enabled. This was then
causing the build to fail since the project extension checks for
generated MSBuild files in the obj directory which are not being
used by .nuproj projects. So instead of checking for a TargetFramework
property in the project the .NET Core project extension checks for
the Sdk attribute on the MSBuildProject instead and is only enabled
if this exists.