[NuGet] Improve detection of .NET Core projects
NuGet packaging projects (.nuproj) that use NuGet.Build.Packaging
0.1.276 NuGet package define a TargetFramework property in an
imported MSBuild .props file. This was then causing the project to
be treated as a .NET Core project by the NuGet addin. This caused
the Dependencies folder to be displayed and the References folder
to be removed. Now the NuGet addin checks the MSBuildProject has the
Sdk attribute instead of looking at the MSBuild properties defined
by the project.