[Packaging] Use package id as description by default for new projects
Fixed bug #52657 - Creating a new Multiplatform Library requires a
description
https://bugzilla.xamarin.com/show_bug.cgi?id=52657

Description is required when creating a NuGet package project or
a multiplatform project. To simplify the project creation process the
package id is used as the description by default. As you type the
package id into the project template wizard page the description text
box will be updated. The description can be changed to be different
to the package id if required.