[DotNetCore] Fix failing test
The test was using a non SDK style project which was originally used
for .NET Core projects until a change was made with later versions of
the .NET Core SDK. The .NET Core addin now checks for the Sdk attribute
existing in the project so the test was failing. The test now uses a
SDK style project.