Fixed 'Bug 55398 - Replace in Files within Current Document gives Value cannot be null error'
Not exactly sure why document.FileName was null. I would've guessed it
was a new empty unsaved file use case - but the screenshots doesn't
look like so. Anyways in the search in open documents use case it's
not needed to look at the file name at all - so that should fix that
issue. But I hadn't any use case for that.