Fixed 'Bug 56806 - Intellisense breaks when typing #region in C# file'
Now with internals access Formatter.GetFormattedTextChanges is
available for us that issue can be fixed. The way creating a new
syntax tree & get the diff to the old tree caused too large text
replace changes. When the caret was in between two changes it moved to
the end of both because the tree diff concated both replace operations
to one. This fixed the bug & on the fly formatting should be way
faster without the work around.