Fix sdk location issue
When creating the local copy of msbuild, copy the whole msbuild bin
directory. The trick of explicitly loading assemblies doesn't work when
msbuild needs to use additional app-domains, and we also need support
files such as Sdks and SdkResolvers. So we now play safe and we
copy the whole dir.