[Ide] Allow New Project dialog to resize to fit text
Fixed bug #54704 - Truncation on dialog "Configure your new Single
View App"
https://bugzilla.xamarin.com/show_bug.cgi?id=54704

With German selected as the IDE language the New Project dialog
would cut off some of the controls on the left hand side of the
final page of the dialog. This was because the dialog had a fixed
width.