[NuGet] Add file template condition for packages
A file template can now specify that it should be enable if the project
has a specific reference or has a particular NuGet package installed.
Whilst a reference will work for project's that use a packages.config
file if the project uses a project.json file or uses PackageReferences
then just checking the references will not find any matches. By
allowing a file template to specify a package id then projects that
do not have an explicit reference to a particular assembly, due to not
using a packages.config file, can now have file templates enabled or
disabled.

<HasPackageOrReference PackageId="Xamarin.Forms" Assembly="Xamarin.Forms" />