[Ide] Fix Remove item not added
Fixed bug #55560 - Setting Build Action to None for C# file is ignored
https://bugzilla.xamarin.com/show_bug.cgi?id=55560

Creating a .NET Core console project, then changing the build action
on a C# file to None, back to Compile and then back to None again
would result in the Remove item not being added to the project file.
So the project file would end up having:

  <ItemGroup>
    <None Include="Program.cs" />
  </ItemGroup>

When it should have had:

  <ItemGroup>
    <Compile Remove="Program.cs" />
  </ItemGroup>
  <ItemGroup>
    <None Include="Program.cs" />
  </ItemGroup>

The problem was that when the Compile Remove item was removed when
switching the build action back to Compile the associated best item
group was not removed since the wrong group id was used. Then when
the new Compile Remove item was added back again on changing the
build action back to None it was added to a item group that did not
belong to the MSBuildProject instead of creating a new item group.