Add a preventative workaround for an incoming MSBuild regression
MSBuild 15.3 will be briefly broken, see details here:
https://github.com/Microsoft/msbuild/issues/2165

It is not yet public and will be fixed soon, but we're already starting to hit the problem as we upgrade. To temporarily work around this until everybody upgrades the MSBuild that fixes the issue this is a simple workaround with no other side effects.