Initial variant to support extern C xammac static registrar
This commit is needed for compliance with Xamarin.Mac changes.

As of:
https://github.com/xamarin/xamarin-macios/commit/51cda3b468b399107459bb63e35221bd5eb89367

We need to check for both C and C++ style exports. Once this
version goes out of stable release / deprecates, we can safely
check for extern "C" unconditionally.

Until then, circumvent this to allow backwards compatibility
by checking whether the C exported symbol exists