[NuGet] Fix credential dialog shown multiple times on opening solution
Fixed bug #56674 - VS for Mac constantly prompting for MyGet
credentials
https://bugzilla.xamarin.com/show_bug.cgi?id=56674

With check for updates enabled, multiple projects in the solution,
and a package source that was missing or had invalid credentials,
on opening the solution the credential dialog would display multiple
times even if the correct username and password was entered or the
dialog was cancelled. The dialog was being displayed for each
project and the credential information was not being re-used.

To fix this the NuGet SourceRepositories are re-used when checking
for updates and also when restoring the projects when the solution
is first opened. Any valid credentials entered will be re-used
when checking the remaining projects. If the credential dialog is
cancelled then the dialog is no longer displayed again whilst
checking for updates.