[NuGet] Fix credential dialog displayed when credentials available
With credentials stored in the key chain the credential dialog would
still be displayed when it should not. The problem was that the
NuGet credential service puts itself in a retry mode if any of the
credential providers are used when trying to authenticate against a
package source. Once in the retry mode then the MonoDevelop credential
provider will always show a dialog asking for the credentials instead
of re-using the existing credentials. To avoid this the credential
service is reset before user actions, such as opening the Add Packages
dialog, running a restore or an update.