[Ide] Custom New Project dialog final page controls use full width
If a custom final page control, such as the Xamarin Test Cloud check
box, does not have an associated information tooltip control then
allow it to extend across the full width on the left hand side of the
New Project dialog. This makes better use of the available space in
case the width of the dialog needs to increase.