Don't crash the main process when a remote builder crashes
I'm investigating https://bugzilla.xamarin.com/show_bug.cgi?id=55660 and I'm noticing a few things that happen when the remote process crashes.

1. Calling Task.ContinueWith with a delegate that throws will crash the entire process.
2. It is happening that while we're iterating over messageWaiters in AbortPendingMessages() a new message is added (presumably reentrancy when the lock is pumping the timer adds a new one on the same thread, so the lock is ineffective). Simple fix is to capture the set of messages to drain and iterate over an immutable snapshot.
3. Several threads are calling SetException on the task source resulting in an InvalidOperationException. Calling TrySetException instead avoids that exception.

Generally, I feel like synchronization in this area needs to be thought about as we should abort more gracefully when the process terminates and avoid trying to contact it again, resulting in a cascade of secondary failures that make things even worse.

This fix is not yet ready because though the IDE doesn't crash anymore, it continues "building" indefinitely and never stops the build.

I need to find a better way to stop in case of exception to signal that the build should stop.