[DotNetCore] Fix ASP.NET Core project build error with old SDK
Fixed bug #57195 - Xamarin.Forms with ASP.NET Mobile Backend does not
build
https://bugzilla.xamarin.com/show_bug.cgi?id=57195

Building an ASP.NET Core project when .NET Core SDK 1.0.3 or lower
was installed, or if no .NET Core SDK was installed, would result in
the build failing with the error:

error MSB4057: The target "Build" does not exist in the project.

With .NET Core SDK 1.0.4 or later installed this build error does not
occur. The problem is that with older .NET Core SDKs the ASP.NET Core
project uses the Microsoft.Web.Sdk.Web MSBuild imports. The older
Microsoft.Web.Sdk.Web  imports use the MSBuildSdksPath property when
importing other MSBuild targets files. This is not defined and results
in nothing being imported and so the build targets are not available.
To fix this the MSBuildSDKsPath property is defined globally if an
old .NET Core sdk is installed, or if no .NET Core SDK is installed.