[Gettext] Do not include CDATA tags in translations
CDATA tags that were inside translatable xml elements were included
in the extracted translation text when the files are scanned. The
CDATA tags themselves should not be displayed to the user so they are
now excluded with the regular expression. CDATA tags are used in
the project templates and were not translated since the text being
passed to the translation catalog would not include these tags and
would not find a match.