[Ide] Fix project template categories not added to translations
The project template categories were defined in the .addin.xml file
using an attribute without an underscore so they were not added
to the translations. The project template categories now use _name
for the category name so they are added to the translations. To
maintain backwards compatibility the name attribute without the
underscore is still supported.

<Category id="library" _name="Library">