[Packaging] Fix description metadata not wrapping in project options
Fixed bug #56918 - When typing a description in NuGet package Metadata
the UI keeps on growing
https://bugzilla.xamarin.com/show_bug.cgi?id=56918

In project options typing a long description for the NuGet package
metadata would cause the dialog to increase in width. The release
notes on the other tab does not have this problem. Setting the
scroll window's policy to never show the horizontal scrolllbars
seems to break the word wrapping. Leaving it as the default fixes
the wrapping so the dialog does not increase in width.