[DotNetCore] Fix .NET Core library project targeting .NET standard
Fixed bug #57591 - .NET Core Class Library targets .NET Standard
https://bugzilla.xamarin.com/show_bug.cgi?id=57591

When a single .NET Core app runtime was installed then the .NET
Core app class library project would use the default framework set
by the project template which is .NET Standard instead of .NET Core
App. This happens when the wizard is not displayed so the framework
is not set. Now the framework is set if the wizard page is not
displayed so the correct target framework is used in the created
class library project.