[NuGet] Ignore ReferenceOutputAssembly false project refs for restore
Project references that have ReferenceOutputAssembly set to false
should not be added to the project.assets.json file. This can cause
the restore to fail. For example, if a .NET Standard project has a
project reference to a .NET Core App project, but has the
ReferenceOutputAssembly set to false, then dotnet restore from the
command line works, but the restore would fail in the IDE. Now the
restore works when run by the IDE.