[Core] Optimize MSBuild evaluation stack allocations and memory retention
This patch reduces the duplicate strings in MD from 90MB to 40MB. The time increase is pretty much unnoticeable - we get rid of StringBuilder.ToString() on duplicate strings, thus we save 1200ms, and we trade that with a search/hash that uses 1900ms. So we're 700ms slower, at the cost of 50MB less managed memory.
There are probably other optimized cases as now the strings are equal by reference, so equals checks would hit fast paths.