[NuGet] Do not open Package Console when NuGet action is cancelled
When the user cancels a NuGet operation from the status bar the
Package Console is now not opened when the NuGet operation fails due
to the cancellation. If the NuGet operation is not cancelled then the
Package Console window will still be opened if the operation fails.
This is also a workaround for the IDE closing the currently displayed
dialog when the Package Console pad opens after the NuGet operation
is cancelled by the user.