[NuGet] Allow actions to be cancelled from background runner
Whilst the background NuGet package actions can be cancelled from the
status bar or the Package Console it is not that simple to cancel
them from elsewhere. The current action is associated with a progress
monitor which has an associated cancellation token source which can
be used by the status bar and the Package console when the monitor
is cancelled. However doing this without having access to the progress
monitor is difficult. So now the background package action runner
provides a Cancel method that can be called to cancel the current
running action and any pending actions.