[NuGet] Allow NuGet action to be cancelled on closing solution
Fixed bug #53795 - Failure to install blocks the IDE
https://bugzilla.xamarin.com/show_bug.cgi?id=53795

Previously when closing the solution or IDE when a NuGet package
action is still running a dialog would be displayed saying that the
solution cannot be closed until the NuGet operation was completed.
This was true with NuGet v2 but with NuGet v3 and above the NuGet
operations can be cancelled. So now the dialog allows the current
operation to be cancelled when the Yes button is clicked. If the
operation is taking a while a busy spinner image will be displayed in
the dialog until the dialog is closed or the operation is cancelled
and then the solution will be closed.