[NuGet] Fix solution closing dialog from shrinking when button clicked
When the No or Yes button on the dialog was clicked the dialog's width would
shrink. Overriding the OnBoundsChanged method and not changing the bounds to
any smaller size did not prevent the dialog being resized. The only way that
seemed to work was to not set Resizable to false for the dialog.

Also switched to using DialogButtons instead of a separate HBox with buttons
so the buttons stay in the bottom right hand corner of the dialog if the dialog
is resized.