[DotNetCore] Show SDK information before dependencies are restored
Fixed bug #56455 - Show SDK for project when not restored
https://bugzilla.xamarin.com/show_bug.cgi?id=56455

When a new .NET Core project is created the Dependencies - SDK folder
shows no items until the restore is completed. Now the SDK dependency
is shown (e.g. NETStandard.Library or Microsoft.NETCore.App) whilst
the dependencies are being restored and the updated information has
not yet been returned from MSBuild. This mirrors the behaviour of
Visual Studio on Windows.