[NuGet] Cancel restore on closing solution without prompting
On closing the solution a dialog will be displayed if an install or
uninstall is currently pending. If a restore is being run then the
background action runner will be asked to cancel it and the solution
will be closed without waiting for the restore to complete. Also
changing the information shown in the dialog indicating whether an
install or an uninstall is currently in progress instead of the
original less specific message about NuGet packages being processed.