[NuGet] Support AssetTargetFallback
Fixed bug #57093 - Add support for AssetTargetFallback
https://bugzilla.xamarin.com/show_bug.cgi?id=57093

.NET Core 2.0 preview 2 uses an AssetTargetFallback MSBuild property
defined in one of the imported SDK MSBuild files. This is used instead
of the PackageTargetFallback property when determining if a NuGet
package is compatible. Currently the AssetTargetFallback property
is set to net461 to allow .NET Core projects to reference NuGet
packages that include assemblies that target the full .NET Framework.
The supported fallback frameworks are now added to the generated
project.assets.json file when a .NET Core 2.0 project is restored.
This then allows .NET Core 2.0 projects to use .NET Framework
NuGet packages.