[Ide] Support not saving files when creating new project
Fixed bug #55351 - Dirty file interferes with creating new project
https://bugzilla.xamarin.com/show_bug.cgi?id=55351

With unsaved files open in the IDE, creating a new project would
show a dialog prompt asking to save changes. This dialog had two
buttons Cancel and Save (or Save As). Now the dialog also has a
Don't Save button which will discard the changes and close the unsaved
file. Also the dialog now has a message saying that not saving will
cause the changes to be permanently lost.