[WebReferences] Disable WCF option in dialog when unsupported
Fixed bug #57449 - Remove WCF 'Add a Reference' dialog
https://bugzilla.xamarin.com/show_bug.cgi?id=57449

The WCF option in the Web References dialog is enabled if the project
targets the .NET Framework. If the project targets a Xamarin mobile
framework then WCF cannot be selected, the combo box is disabled, and
instead the .NET 2.0 Web Services option is selected by default. The
logic for PCL projects is unchanged from before and WCF will selected
by default and the .NET 2.0 Web Services option cannot be selected.