[A11y] Fix the keyboard focus on the editor
Now that the whole of the shell UI can be focused, lots of controls like to
steal the keyboard focus meaning that when we expect to be able to type in the
editor, we often can't. To fix this, when opening a new document, we have the
editor grab focus.