[Core] Provide API to get referenced projects with any aliases.
The existing GetReferencedAssemblyProjects returns the DotNetProjects
that are directly referenced without the alias information. The
type system service also requires the associated aliases for the
referenced project and so previously was determining them separately
based on the directly referenced projects. This prevents the
DotNetProject from returning transitively referenced projects from the
GetReferencedAssemblyProjects method. So a new method has been added
GetReferencedAssemblyProjectAliases which returns the referenced
projects and any associated aliases. This will allow the DotNetProject
to return transitively referenced projects and the type system can
just use this information without having to lookup the aliases
separately and missing projects that are not directly referenced.