[NuGet] Fix null ref when opening NuGet dialog from code action
A null reference exception would occur when the solution was
selected and the Install with package manager code action was
selected. This was because there was no active project selected in
the IDE which was not handled. Now the package services gets the
project from the active document before showing the dialog which
prevent a null reference exception from being thrown.