Fix issue in FastCheckNeedsBuild
The FastCheckNeedsBuild method now takes an OperationContext as argument.
This is necessary because when building a project, OperationContext can
contain custom msbuild properties that can have an effect on the build
(for example, a build can be specific to the target device currently selected
in the IDE).

OnFastCheckNeedsBuild now keeps track of the global msbuild properties
specified in the context, and uses them when checking for changes.