[NuGet] Fix transitive assembly references not available until restart
Fixed bug #58009 - Transitive assembly references not available until
project reloaded
https://bugzilla.xamarin.com/show_bug.cgi?id=58009

With a solution containing three .NET Standard projects: LibC
references LibB which references LibA. If the Newtonsoft.Json NuGet
package was installed into LibA the types from this NuGet packages
are not available in LibB or LibC until the solution is closed and
re-opened again. Closing and re-opening the solution refreshes the
reference information for the type system. To fix this when a NuGet
package is installed into a .NET Core project the projects that
reference this project have their reference information refreshed
by calling NotifyModified ("References") which causes the type system
to get the reference information again from MSBuild. Types from
the installed NuGet packages are then available in projects that
reference this updated project either directly or indirectly.