[Ide] Fix project not created in own directory with New Project dialog
Fixed bug #57934 - When adding project to solution "Create a project
directory within the solution directory." checkbox is not respected
https://bugzilla.xamarin.com/show_bug.cgi?id=57934

To reproduce the original bug:

1. Create a new solution using the New Project dialog and ensure
the Create a project directory within the solution directory is
unchecked.
2. Add a new project to the existing solution.
3. The final page of the New Project dialog would have the Create
a project directory within the solution directory check box checked
and disabled. However the project would not be created in its own
directory.