[Ide] Support selecting multiple rows in Errors pad
Fixed bug #11946 - Need a way to export build errors and warnings
https://bugzilla.xamarin.com/show_bug.cgi?id=11946

Allow multiple rows to be selected in the Errors pad. The Copy menu
will be enabled if multiple rows are selected. The Go to and Show
Error Reference menus will be disabled if multiple rows are selected.
The Copy menu will copy all selected rows to the clipboard.