[F#] Fix Android project template compile tests
The tests that create Android projects and compile them were failing
with 'The OutputPath property is not set for project'. This was
because the projects were saved after NuGet packages were installed.
The new Android projects use the Xamarin.Build.Download NuGet package
which adds a .props Import at the start of the project file. Saving
the Android project after this NuGet package was added would result
in the Xamarin.Android.FSharp.targets Import being added at the top
of the project file, since similar elements are grouped together.
The Xamarin.Android.FSharp.targets Import needs to be at the end of
the project file. Saving the project before adding the NuGet
packages fixes this problem.