[Ide] Fix failure to create CocoaApp project from New Project dialog
Fixed bug #58412 - Unable to create CocoaApp from recent templates
https://bugzilla.xamarin.com/show_bug.cgi?id=58412

After creating a CocoaApp (Mac - App - General) that targets
Mavericks if the CocoaApp project template was selected from the
recent projects list in the New Project dialog the project would not
be created if a later Mac OS was selected. In the IDE log an error
would be reported:

No template found matching condition 'Yosemite=false'.

The problem was that the selected project template was part of a
group but the group was not selected. When a different Mac OS was
selected for the new project the condition applied would result in
no template being found.