[Ide] Ensure New Project dialog language selector matches language
Fixed bug #58392 - Blank Forms App Template Uses C# Language Instead
of Previous Selection
https://bugzilla.xamarin.com/show_bug.cgi?id=58392

The language button selector was not updated when a recently used
project template was selected on opening the New Project dialog. This
resulted in the selected language in the UI not being the language
used when creating the project. Now the language selector is updated
if the language is changed when a recently used project template is
selected.