[DotNetCore] Prevent IDE being terminated after debugging unit tests
With the latest .NET Core 2.0 SDK installed when a .NET Core test
project is debugged the IDE would be terminated after the debugging
session had finished. The problem was that currently the process id
sent to VSTest is the IDE process and this is terminated by the test
runner in the latest recent versions. A workaround here is to return a
process id of 1. .NET Core 1.0.4 SDK does not work if a process id
of 1 is returned so the old behaviour remains if the latest SDK is
lower than version 2.0. The real fix is to return the correct process
id for the application being debugged. Until that is fixed these
workarounds will remain.