[UnitTests] Make recent template tests more robust
The RecentTemplates TwoTemplatesInGroupDifferentLanguage
TreatedAsDifferentRecentTemplate test occasionally fails to find any
recent templates. To make the recent template tests more robust the
list of templates is retried several times before failing.
This handles the case where the RecentFileStorage's constructor
occasionally clears out the cached in-memory list whilst the test is
adding new items. The in-memory list is eventually updated when the
file is saved but this can take some time to happen since the save
is delayed.

Also a different file name is used with the recent file storage
for each unit test to minimize the chance of a sharing violation
from occuring.