[Ide] Trim command and working directory for external tools
Spaces at the start or end of the Command or Working directory text
boxes when creating an External Tool will result in the tool failing
to run. It is not obvious what the problem is.

With spaces in the Command you see an error message saying that the
External program execution failed and you cannot easily tell there is
an extra space causing the problem.

With spaces at the start or end of the Working directory when
the external tool is being run with Mono you see an error "Process
'mono' has exited with error code 255" in the Application Output
window.

The Command and Working directory text is now trimmed when the
External Tool is created to avoid the above errors from occuring.