[Ide] Support backslashes in template primary output paths
Fixed bug #58980 - Back slashes in .NET project templates result in
project not being added to solution
https://bugzilla.xamarin.com/show_bug.cgi?id=58980

The template.json file used by the new templating engine uses a
primaryOutputs property for projects and files. If the paths used
contained backslashes then the project would not be added to the
solution.

"primaryOutputs": [
  { "path": "Library1\\Library1.csproj" },
  { "path": "Library2\\Library2.csproj" }
]

Now the paths have the backslashes replaced with forward slashes if
the current platform does not support backslash as a directory
separator.