[Core] Add initial support for unifying project and assembly references
DotNetProject now has a GetReferences method that returns project and
assembly references. The assemblies are returned from the existing
GetReferencedAssemblies method and the project references are taken from
the project. In the future both project and assembly references will be
taken from MSBuild. The GetReferences and associated project extension
methods allows the .NET Core project extension to provide transitive
project references. In the future this will be information will be
available from MSBuild.