Update to x.y.3.z for 15.3 SR3
Reason for Patch: 

```
Release communications consistency.  We had to include the full build number in the release notes sections for the 15.3.2 release because the all of the builds were still versioned as x.x.0.x.  We can strive to improve upon that for the 15.3.3 release by aligning the 3rd-digit version number with the Visual Studio 2017 service release number.
```