[Core] Re-evaluate AppliesTo condition on capability change
If the project extension uses an AppliesTo attribute which is
initially false and then a new project capability is added and the
project is re-evaluated then the AppliesTo attribute is now
re-evaluated. This will then cause the project extension to be added
to the extension chain. If the project extension was attached to the
project and the project capability is removed then the project
extension will now be removed when the project is re-evaluated.