[Core] Fix new project being added with ToolsVersion 15.0
The default ToolsVersion for projects is 4.0. When adding a new
project to an existing solution the new project would have 15.0 used
for its ToolsVersion. The default tools version used for a project
is determined by the solution's file format. Since the VS 2017 file
format was the first returned it is used by default in some cases
and would cause its default tools version of 15.0 to be used. To
fix this the VS 2012 file format is returned first so it will be
used by default for solution file format version 12.0.