[NuGet] Support RestoreProjectStyle MSBuild property
Fixed bug #58985 - Support RestoreProjectStyle MSBuild property
https://bugzilla.xamarin.com/show_bug.cgi?id=58985

If a project sets the RestoreProjectStyle property to be
PackageReference then it should be restored as though it has
PackageReferences even if it does not have any.

 <RestoreProjectStyle>PackageReference</RestoreProjectStyle>

This fixes runtime problems when a project references a .NET Standard
project that has PackageReferences since without this assemblies
from the NuGet packages are not copied to the output directory and
would have to be added to the main project.